fix(e2e): resolve qBound<qint64> overload ambiguity on GCC in play_harness
The stress scenario's qBound<qint64>(3500, durMs / 5, qint64(6000)) mixed int,
long, and qint64 arguments. Clang accepted it, but GCC (the Linux CI compiler)
rejected the call as ambiguous under -Werror, breaking the Linux build. Make all
three arguments qint64 so the overload resolves unambiguously on both compilers.

+++ tests/e2e/play_harness.cpp
@@ -1403,7 +1403,8 @@ int main(int argc, char** argv) {
                             (long long) err);
                 });
             mon->start(16);
-            const int stressFinishMs = int(qBound<qint64>(3500, durMs / 5, qint64(6000)));
+            const int stressFinishMs =
+                int(qBound<qint64>(qint64(3500), qint64(durMs) / 5, qint64(6000)));
             QTimer::singleShot(stressFinishMs, &app, finish);
 
         } else if (scen == "devicelost") {
